TEST: fix failed testers and limit `testpaths`

+++ tests/test_image_registration_tool.py
@@ -35,7 +35,7 @@ class TestImageRegistrationTool(tutils.BaseTester):
         )
         
         acquisition_tool = SimulatedAcquireImage(whole_image, return_message=False)
-        registration_tool = ImageRegistration()
+        registration_tool = ImageRegistration(registration_method="phase_correlation")
         
         acquisition_tool.acquire_image(
             y_center=184, x_center=184, size_y=128, size_x=128
@@ -71,7 +71,10 @@ class TestImageRegistrationTool(tutils.BaseTester):
         )
         
         acquisition_tool = SimulatedAcquireImage(whole_image, return_message=False)
-        registration_tool = ImageRegistration(image_coordinates_origin="top_left")
+        registration_tool = ImageRegistration(
+            image_coordinates_origin="top_left",
+            registration_method="phase_correlation",
+        )
         
         acquisition_tool.acquire_image(
             y_center=184, x_center=184, size_y=128, size_x=128
@@ -133,6 +136,43 @@ class TestImageRegistrationTool(tutils.BaseTester):
         assert np.allclose(offset, [-20, -20], atol=1.0)
         return
 
+    def test_image_registration_ncc(self):
+        np.random.seed(123)
+
+        whole_image = tifffile.imread(
+            os.path.join(
+                self.get_ci_input_data_dir(),
+                'simulated_images',
+                'cameraman.tiff'
+            )
+        )
+
+        acquisition_tool = SimulatedAcquireImage(whole_image, return_message=False)
+        registration_tool = ImageRegistration(
+            registration_method="ncc",
+            registration_algorithm_kwargs={"max_shift": 25},
+        )
+
+        acquisition_tool.acquire_image(
+            y_center=184, x_center=184, size_y=128, size_x=128
+        )
+        acquisition_tool.acquire_image(
+            y_center=164, x_center=164, size_y=128, size_x=128
+        )
+
+        previous_info = acquisition_tool.get_previous_image_info()
+        current_info = acquisition_tool.get_current_image_info()
+        offset = registration_tool.register_images(
+            image_t=acquisition_tool.image_k,
+            image_r=acquisition_tool.image_km1,
+            psize_t=current_info["psize"],
+            psize_r=previous_info["psize"],
+            registration_algorithm_kwargs={"max_shift": 25},
+        )
+
+        assert np.allclose(offset, [-20, -20])
+        return
+
     def test_image_registration_from_buffer_arrays(self):
         np.random.seed(123)
 
@@ -145,18 +185,28 @@ class TestImageRegistrationTool(tutils.BaseTester):
         )
 
         acquisition_tool = SimulatedAcquireImage(whole_image, return_message=True)
-        registration_tool = ImageRegistration()
+        registration_tool = ImageRegistration(registration_method="phase_correlation")
 
-        acquisition_tool.acquire_image(
+        acquisition_results = []
+        acquisition_results.append(acquisition_tool.acquire_image(
             y_center=184, x_center=184, size_y=128, size_x=128
-        )
-        acquisition_tool.acquire_image(
+        ))
+        acquisition_results.append(acquisition_tool.acquire_image(
             y_center=164, x_center=164, size_y=128, size_x=128
-        )
+        ))
 
-        offset = self.register_previous_current(acquisition_tool, registration_tool)
+        try:
+            offset = self.register_previous_current(acquisition_tool, registration_tool)
 
-        assert np.allclose(offset, [-20, -20])
+            assert np.allclose(offset, [-20, -20])
+        finally:
+            for acquisition_result in acquisition_results:
+                for path in (
+                    acquisition_result["img_path"],
+                    acquisition_result["raw_data_path"],
+                ):
+                    if os.path.exists(path):
+                        os.remove(path)
         return
 
     def test_image_registration_from_dumped_array_paths(self):
@@ -171,20 +221,23 @@ class TestImageRegistrationTool(tutils.BaseTester):
         )
 
         acquisition_tool = SimulatedAcquireImage(whole_image, return_message=True)
-        registration_tool = ImageRegistration()
+        registration_tool = ImageRegistration(registration_method="phase_correlation")
 
-        acquisition_tool.acquire_image(
+        acquisition_results = []
+        acquisition_results.append(acquisition_tool.acquire_image(
             y_center=184, x_center=184, size_y=128, size_x=128
-        )
-        acquisition_tool.acquire_image(
+        ))
+        acquisition_results.append(acquisition_tool.acquire_image(
             y_center=164, x_center=164, size_y=128, size_x=128
-        )
-        current_info = acquisition_tool.get_current_image_info()
-        previous_info = acquisition_tool.get_previous_image_info()
-        current_path = acquisition_tool.dump_array("image_k")["array_path"]
-        previous_path = acquisition_tool.dump_array("image_km1")["array_path"]
+        ))
+        current_path = None
+        previous_path = None
 
         try:
+            current_info = acquisition_tool.get_current_image_info()
+            previous_info = acquisition_tool.get_previous_image_info()
+            current_path = acquisition_tool.dump_array("image_k")["array_path"]
+            previous_path = acquisition_tool.dump_array("image_km1")["array_path"]
             offset = registration_tool.get_offset_from_paths(
                 current_image_path=current_path,
                 reference_image_path=previous_path,
@@ -195,7 +248,15 @@ class TestImageRegistrationTool(tutils.BaseTester):
             assert np.allclose(offset, [-20, -20])
         finally:
             for array_path in (current_path, previous_path):
-                os.remove(array_path)
+                if array_path is not None and os.path.exists(array_path):
+                    os.remove(array_path)
+            for acquisition_result in acquisition_results:
+                for path in (
+                    acquisition_result["img_path"],
+                    acquisition_result["raw_data_path"],
+                ):
+                    if os.path.exists(path):
+                        os.remove(path)
         return
